safari-534.53-branch/LayoutTests/fast/dom/Geolocation/timeout-zero-expected.txt

Tests that when timeout is zero (and maximumAge is too), the error callback is called immediately with code TIMEOUT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.TIMEOUT
PASS error.message is "Timeout expired"
PASS successfullyParsed is true

TEST COMPLETE
